requirements.txt
pypdf
textacy
elasticsearch[async]
httpx
python-dotenv
fastapi
uvicorn